+++ CHANGES.md
@@ -6,6 +6,7 @@
 ### Fixed
 
 * vertigo-cli: `--mount-point` option fixed (broken by hydration)
+* Nested `Computed` recomputations [#472]
 
 ## 0.9.0 - 2025-12-01